UnzipBlobFunction/requirements.txt
azure-functions
azure-storage-blob
gzip